add ipv6 parsing support
