+++ firewall/CMakeLists.txt
@@ -10,6 +10,7 @@ set(SRC
     ./src/packet-eth.cc
     ./src/packet-arp.cc
     ./src/packet-ipv4.cc
+    ./src/packet-ipv6.cc
     ./src/packet-icmp.cc
     ./src/firewall_event_mgr.cc)
