+++ firewall/include/event_types.h
@@ -16,6 +16,7 @@ enum event_type {
     IPV4_DONT_FRAGMENT_SET,
     IPV4_MORE_FRAGMENT_SET,
     IPV4_MF_DF_FLAGS_SET,
+    IPV6_VERSION_INVAL,
     UNSUPPORTED_ETHERTYPE,
     UNSUPPORTED_PROTOCOL,
     PACKET_LEN_TOO_SHORT,
