+++ firewall/include/packet.h
@@ -7,12 +7,15 @@
 #include <event_types.h>
 #include <packet-buf.h>
 #include <packet-eth.h>
+#include <packet-arp.h>
+#include <packet-ipv4.h>
 #include <packet-icmp.h>
 
 #define PACKET_BUF_MAX_DATA_LEN 8192
 
 #define ETHERTYPE_ARP       0x0806
 #define ETHERTYPE_IPV4      0x0800
+#define ETHERTYPE_IPV6      0x86DD
 
 namespace nos::firewall {
 
@@ -29,72 +32,6 @@ struct ieee8021ae_header {
 struct ieee8021x_header {
     event_type deserialize(packet_buf &buf);
     void free_hdr() { }
-};
-
-#define ARP_OP_ARP_REQ          1
-#define ARP_OP_ARP_REPLY        2
-#define ARP_OP_RARP_REQ         3
-#define ARP_OP_RARP_REPLY       4
-#define ARP_OP_DRARP_REQ        5
-#define ARP_OP_DRARP_REPLY      6
-#define ARP_OP_DRARP_ERROR      7
-#define ARP_OP_INARP_REQ        8
-#define ARP_OP_INARP_REPLY      9
-
-#define ARP_HEADER_LEN          28
-
-/**
- * @brief - Implements ARP header.
- */
-struct arp_header {
-    uint16_t        header_type;
-    uint16_t        protocol_type;
-    uint8_t         hwaddr_len;
-    uint8_t         protoaddr_len;
-    uint16_t        operation;
-    uint8_t         sender_hwaddr[MACADDR_LEN];
-    uint32_t        sender_proto_addr;
-    uint8_t         target_hwaddr[MACADDR_LEN];
-    uint32_t        target_proto_addr;
-
-    inline bool is_arp_req() { return operation == ARP_OP_ARP_REQ; }
-    inline bool is_arp_reply() { return operation == ARP_OP_ARP_REPLY; }
-    inline bool is_proto_ipv4() { return protocol_type == UNSUPPORTED_ETHERTYPE; }
-
-    event_type deserialize(packet_buf &buf);
-    void free_hdr() { }
-};
-
-#define IPV4_VERSION 4
-#define IPV4_IHL_LEN_MIN 5
-#define IPV4_IHL_LEN_MAX 15
-
-/**
- * @brief - Implements ipv4 header.
- */
-struct ipv4_header {
-    uint8_t         version;
-    uint8_t         ihl;
-    uint8_t         dscp;
-    uint8_t         ecn;
-    uint16_t        total_len;
-    uint16_t        id;
-    uint8_t         flags_reserved:1;
-    uint8_t         flags_dont_fragment:1;
-    uint8_t         flags_more_fragment:1;
-    uint16_t        frag_off;
-    uint8_t         ttl;
-    uint8_t         protocol;
-    uint16_t        hdr_chksum;
-    uint32_t        source_ipaddr;
-    uint32_t        dest_ipaddr;
-    bool            header_parsed_ok;
-
-    event_type deserialize(packet_buf &buf);
-    void free_hdr() { }
-
-    inline bool has_fragment()
-    { return (flags_more_fragment == 1) || (frag_off != 0); }
 };
 
 struct ipv6_header {
