+++ firewall/src/firewall-parser.cc
@@ -33,6 +33,12 @@ event_type firewall_intf::parse_packet(packet_parser_state &state)
                 type = parse_protocol(state);
             }
         break;
+        case ETHERTYPE_IPV6:
+            type = state.pkt.ipv6_h.deserialize(state.pkt_buf);
+            if (type == event_type::NO_ERROR) {
+                type = parse_protocol(state);
+            }
+        break;
         default:
             return event_type::UNSUPPORTED_ETHERTYPE;
     }
